Fixed include

+++ src/main.c
@@ -1,4 +1,8 @@
+#include <arpa/inet.h>
 #include <pthread.h>
+#include <stdlib.h>
+#include <sys/socket.h>
+#include <unistd.h>
 
 #include "request.h"
 
@@ -9,7 +13,7 @@ int main(int argc, char *argv[])
     unsigned int clnt_adr_size;
     pthread_t t_id;
 
-    uint16_t server_port = 8080;
+    unsigned int server_port = 8080;
     if (argc > 1) {
         if (sscanf(argv[1], "%u", &server_port) == 0 || server_port > 65535) {
             fprintf(stderr, "error: invalid command line argument, using default port 8080.\n");
